refactor(ui): extract session turn changes panel
